Rename build_configuration to configuration
